- Move to utils.py - Move const to const.py
